requirements.txt
tweepy
google-genai
python-dotenv
